refactor and clean code
